pyproject.toml
[tool.poetry]
name = "oscal_clpy"
version = "0.1.0"
description = "A wrapper for the oscal-cli Java utility, like tabula-py to process PDFs with tabula."
authors = ["Al S <xee5ch@member.fsf.org>"]
license = "GPL-3.0"

[tool.poetry.dependencies]
python = "^3.10"
JPype1 = "^1.4.1"

[tool.poetry.dev-dependencies]
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
